Write functions that orders ranges into a tree
This is the first step towards building a tree representation of the
entirety of the formatted text. This tree just contains information
about ranges, but not the underlying text.

There's a test for the current part of the work. That test is likely to
evolve into something more high level as this work continues.