Remove implicit constraint for ranges
In the AllenAlgebra code we were implicitly making this assumption:

    range.start <= range.end

This assumption held for the way we were using the algebra, but it
s still better not to have the assumption in the first place.

We don't need to make this assumption at all, if we take `start` and
`end` to be two bounds of a range, and manually check which one is the
lower bound and which one the upper.